+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -298,7 +298,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-21T08:05:43.551Z",
+    "2025-10-21T10:57:44.779Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -338,7 +338,7 @@ export const UpdateDatasetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-21T08:05:43.551Z"))
+  updated: z.date().default(() => new Date("2025-10-21T10:57:44.779Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {