+++ packages/orq-rc/RELEASES.md
@@ -4409,3 +4409,13 @@ Based on:
 - [typescript v3.14.0-rc.41] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.41] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.41 - packages/orq-rc
+
+## 2025-10-21 10:59:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.637.3 (2.727.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.42] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.42] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.42 - packages/orq-rc